requirements.txt
streamlit
qrcode[pil]
opencv-python
numpy
pillow
